Refactor to use cleaner auth implementation
